move schunk driver to utils
